+++ packages/article-list/article-list.stories.js
@@ -1,0 +1,4 @@
+import { showcaseConverter } from "@times-components/storybook";
+import showcase from "./article-list.showcase";
+
+showcaseConverter(module, showcase);