+++ packages/author-profile/src/author-profile-list-item-prop-types.js
@@ -1,9 +1,0 @@
-import PropTypes from "prop-types";
-
-export const propTypes = {
-  item: PropTypes.object
-};
-
-export const defaultProps = {
-  item: {}
-};