+++ packages/author-profile/src/author-profile-tracking-context.js
@@ -4,7 +4,7 @@ import { withTrackingContext } from "@times-components/tracking";
 export default Component =>
   withTrackingContext(Component, {
     trackingObjectName: "AuthorProfile",
-    getAttrs: ({ author, page, pageSize } = {}) => ({
+    getAttrs: ({ author, page, pageSize }) => ({
       authorName: author && author.name,
       articlesCount: get(author, "articles.count", 0),
       page,