+++ packages/article-flag/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.10"></a>
+## [0.12.10](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.12.9...@times-components/article-flag@0.12.10) (2018-05-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-flag
+
 <a name="0.12.9"></a>
 ## [0.12.9](https://github.com/newsuk/times-components/compare/@times-components/article-flag@0.12.8...@times-components/article-flag@0.12.9) (2018-05-08)