+++ packages/brightcove-video/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.2.11"></a>
+## [2.2.11](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@2.2.10...@times-components/brightcove-video@2.2.11) (2018-05-10)
+
+
+
+
+**Note:** Version bump only for package @times-components/brightcove-video
+
 <a name="2.2.10"></a>
 ## [2.2.10](https://github.com/newsuk/times-components/compare/@times-components/brightcove-video@2.2.9...@times-components/brightcove-video@2.2.10) (2018-05-08)